Move basic time tools into separate module
